* Warnings-- in libvorbiscomment
git-svn-id: https://svn.xiph.org/trunk/oggdsf@8667 0101bb08-14d6-0310-b084-bc0e0c8e3800

+++ src/lib/core/ogg/libVorbisComment/SingleVorbisComment.cpp
@@ -66,7 +66,7 @@ string SingleVorbisComment::toString() {
 
 unsigned long SingleVorbisComment::length() {
 	//FIX::: This would be faster to add them manually.
-	return toString().length();
+	return (unsigned long)toString().length();
 }
 
 bool SingleVorbisComment::parseComment(string inCommentString) {
